Summary: Fixed signed comparison warnings
svn path=/software/LOOS/trunk/; revision=15137

+++ AtomicGroup.cpp
@@ -1012,7 +1012,7 @@ namespace loos {
 
 
   void AtomicGroup::setCoords(double* seq, int m, int n) {
-    if (n != 3 || m != size())
+    if (n != 3 || static_cast<uint>(m) != size())
       throw(LOOSError("Invalid dimensions in AtomicGroup::setCoords()"));
 
     for (int j=0; j<m; ++j)
@@ -1023,8 +1023,8 @@ namespace loos {
 
   void AtomicGroup::getCoords(double** outseq, int* m, int* n) {
     double* dp = static_cast<double*>(malloc(size() * 3 * sizeof(double)));
-    for (int j=0; j<size(); ++j)
-      for (int i=0; i<3; ++i)
+    for (uint j=0; j<size(); ++j)
+      for (uint i=0; i<3; ++i)
 	dp[j*3+i] = atoms[j]->coords()[i];
 
     *m = size();
